Split database provider in order to have one file per provider
